requirements.txt
fastapi==0.100.0        
uvicorn==0.22.0           
pydantic==1.10.7          
faiss-cpu==1.7.3          
numpy==1.24.3           
sentence-transformers==2.2.2
scikit-learn==1.2.2     
psutil